Refactor lowerer feature overrides into partials
