+++ src/Raven.CodeAnalysis/BoundTree/Lowering/Lowerer.Blocks.cs
@@ -1,0 +1,48 @@
+using System.Collections.Generic;
+
+namespace Raven.CodeAnalysis;
+
+internal sealed partial class Lowerer
+{
+    public override BoundNode? VisitBlockStatement(BoundBlockStatement node)
+    {
+        var statements = new List<BoundStatement>();
+
+        foreach (var statement in node.Statements)
+        {
+            statements.Add((BoundStatement)VisitStatement(statement));
+        }
+
+        return new BoundBlockStatement(statements, node.LocalsToDispose);
+    }
+
+    public override BoundNode? VisitIfStatement(BoundIfStatement node)
+    {
+        var condition = (BoundExpression)VisitExpression(node.Condition)!;
+        var thenStatement = (BoundStatement)VisitStatement(node.ThenNode);
+        var elseStatement = node.ElseNode is null ? null : (BoundStatement)VisitStatement(node.ElseNode);
+
+        if (elseStatement is null)
+        {
+            var endLabel = CreateLabel("if_end");
+            return new BoundBlockStatement([
+                new BoundConditionalGotoStatement(endLabel, condition, jumpIfTrue: false),
+                thenStatement,
+                CreateLabelStatement(endLabel),
+            ]);
+        }
+        else
+        {
+            var elseLabel = CreateLabel("if_else");
+            var endLabel = CreateLabel("if_end");
+            return new BoundBlockStatement([
+                new BoundConditionalGotoStatement(elseLabel, condition, jumpIfTrue: false),
+                thenStatement,
+                new BoundGotoStatement(endLabel),
+                new BoundLabeledStatement(elseLabel, elseStatement),
+                CreateLabelStatement(endLabel),
+            ]);
+        }
+    }
+}
+
